WSTEAM1-240: Split MaskedImage into seperate directory
